feat: add the compute_trajectory function.
